fix overflow

+++ experiment/benchmark_arithmetic_decoding.cpp
@@ -13,16 +13,16 @@ int main() {
     cout << std::ctime(&timeinfo) << endl;
 
     std::mt19937_64 engine(0);
-    auto random_arith = [&](int tree_n) -> BitArray {
-        std::vector<uint32_t> left_seq;
+    auto random_arith = [&](int64_t tree_n) -> BitArray {
+        std::vector<uint64_t> left_seq;
         left_seq.reserve(tree_n);
-        std::queue<uint32_t> subtree_size_st;
+        std::queue<uint64_t> subtree_size_st;
         subtree_size_st.push(tree_n);
-        for (int i = 0; i < tree_n; i++) {
-            uint32_t cur_size = subtree_size_st.front();
+        for (int64_t i = 0; i < tree_n; i++) {
+            uint64_t cur_size = subtree_size_st.front();
             subtree_size_st.pop();
-            uint32_t left_size = engine() % cur_size;
-            uint32_t right_size = cur_size - left_size - 1;
+            uint64_t left_size = engine() % cur_size;
+            uint64_t right_size = cur_size - left_size - 1;
             left_seq.push_back(left_size);
             if (left_size) {
                 subtree_size_st.push(left_size);
@@ -34,11 +34,11 @@ int main() {
         return left_seq_to_arithmetic<false>(left_seq);
     };
 
-    for (int s = 1; s <= 20; s++) {
-        int n = 1 << s;
-        int Q = 1 << (25 - s);
+    for (int64_t s = 1; s <= 20; s++) {
+        int64_t n = 1 << s;
+        int64_t Q = 1 << (25 - s);
         vector<BitArray> arith_list;
-        for (int i = 0; i < Q; i++) {
+        for (int64_t i = 0; i < Q; i++) {
             arith_list.push_back(random_arith(n));
         }
