+++ experiment/benchmark_hypersuccinct_binary_trees.cpp
@@ -6,7 +6,8 @@ using namespace std;
 using namespace hyperrmq;
 
 void output_benchmark(
-    pair<vector<pair<string, uint64_t>>, vector<pair<string, int>>> bm_result) {
+    pair<vector<pair<string, uint64_t>>, vector<pair<string, int64_t>>>
+        bm_result) {
     auto [memory_result, time_result] = bm_result;
     cout << "memory:" << endl;
     for (auto&& [label, val] : memory_result) {
@@ -19,7 +20,7 @@ void output_benchmark(
     }
 }
 
-template <int N, int Q, int B, int W,
+template <int64_t N, int64_t Q, int64_t B, int64_t W,
           typename CompressedMicrotreeSplitRankArray>
 void benchmark() {
     using Node = typename HypersuccinctBinaryTree<
@@ -34,9 +35,9 @@ void benchmark() {
     cout << "W = " << W << endl;
     cout << typeid(CompressedMicrotreeSplitRankArray).name() << endl;
 
-    vector<pair<string, int>> time_result;
+    vector<pair<string, int64_t>> time_result;
     auto start = chrono::system_clock::now();
-    auto timer = [&]() -> int {
+    auto timer = [&]() -> int64_t {
         auto now = chrono::system_clock::now();
         return static_cast<double>(
             chrono::duration_cast<chrono::microseconds>(now - start).count() /
@@ -51,7 +52,7 @@ void benchmark() {
     };
 
     mt19937 engine(0);
-    vector<int32_t> perm(N);
+    vector<int64_t> perm(N);
     iota(perm.begin(), perm.end(), 0);
     shuffle(perm.begin(), perm.end(), engine);
 
@@ -60,87 +61,87 @@ void benchmark() {
         cartesian_tree_bp(perm), B);
     time_keeper("cnst.", 1);
 
-    std::vector<int> inorders(Q + 1);
-    for (int i = 0; i < Q + 1; i++) {
+    std::vector<int64_t> inorders(Q + 1);
+    for (int64_t i = 0; i < Q + 1; i++) {
         inorders[i] = engine() % N;
     }
 
     std::vector<Node> nodes;
     time_keeper("to node", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         nodes.push_back(hbt.inorder_to_node(inorders[i]));
     }
     time_keeper("to node", 1);
     nodes.push_back(hbt.inorder_to_node(inorders.back()));
 
     time_keeper("to in.", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.node_to_inorder(nodes[i]);
     }
     time_keeper("to in.", 1);
 
     time_keeper("root", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.root();
     }
     time_keeper("root", 1);
 
     time_keeper("parent", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.parent(nodes[i]);
     }
     time_keeper("parent", 1);
 
     time_keeper("left", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.left_child(nodes[i]);
     }
     time_keeper("left", 1);
 
     time_keeper("right", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.right_child(nodes[i]);
     }
     time_keeper("right", 1);
 
     time_keeper("is leaf", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.is_leaf(nodes[i]);
     }
     time_keeper("is leaf", 1);
 
     time_keeper("child_label", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.child_label(nodes[i]);
     }
     time_keeper("child_label", 1);
 
     time_keeper("subtree", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.subtree_size(nodes[i]);
     }
     time_keeper("subtree", 1);
 
     time_keeper("is anc.", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.is_ancestor(nodes[i], nodes[i + 1]);
     }
     time_keeper("is anc.", 1);
 
     time_keeper("ldesc.", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.leftmost_desc(nodes[i]);
     }
     time_keeper("ldesc.", 1);
 
     time_keeper("rdesc.", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.rightmost_desc(nodes[i]);
     }
     time_keeper("rdesc.", 1);
 
     time_keeper("lca", 0);
-    for (int i = 0; i < Q; i++) {
+    for (int64_t i = 0; i < Q; i++) {
         hbt.lca(nodes[i], nodes[i + 1]);
     }
     time_keeper("lca", 1);
@@ -162,7 +163,7 @@ void benchmark() {
 int main() {
     cout << fixed << setprecision(5);
 
-    constexpr int N = 1e9, Q = 1e6;
+    constexpr int64_t N = 1e9, Q = 1e6;
 
     benchmark<N, Q, 8, 16, CompressedMicrotreeSplitRankArrayHuffman<16>>();
     benchmark<N, Q, 64, 16,
